Extract previewable option to components
