+++ settings-boundary/index.ts
@@ -4,3 +4,4 @@ export * from "./AppearanceSettings"
 export * from "./AccountInfo"
 export * from "./GeneralSettings"
 export * from "./BlockList"
+export * from "./CalendarSettings"
